move import of openquake PshaAdapter one module lower to avoid circular import
