+++ nzshm_model/model.py
@@ -3,7 +3,8 @@ import warnings
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-from nzshm_model.psha_adapter import NrmlDocument, OpenquakeSimplePshaAdapter
+# from nzshm_model.psha_adapter import NrmlDocument, OpenquakeSimplePshaAdapter
+from nzshm_model.psha_adapter.openquake import NrmlDocument, OpenquakeSimplePshaAdapter
 from nzshm_model.source_logic_tree import SourceLogicTree, SourceLogicTreeV1
 
 if TYPE_CHECKING:
