+++ nzshm_model/psha_adapter/__init__.py
@@ -1,2 +1,2 @@
-from .openquake import NrmlDocument, OpenquakeSimplePshaAdapter
+# from .openquake import NrmlDocument, OpenquakeSimplePshaAdapter
 from .psha_adapter_interface import PshaAdapterInterface
