+++ tests/psha_adapter/openquake/test_logic_tree_from_model.py
@@ -5,7 +5,7 @@ from pathlib import Path, PurePath
 import pytest
 
 import nzshm_model
-from nzshm_model.psha_adapter import OpenquakeSimplePshaAdapter
+from nzshm_model.psha_adapter.openquake import OpenquakeSimplePshaAdapter
 
 MODEL = nzshm_model.get_model_version('NSHM_v1.0.4')
 FIXTURE_PATH = Path(__file__).parent / "fixtures"
