+++ tests/psha_adapter/openquake/test_logic_tree_to_simple_nrml.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 import pytest
 
-from nzshm_model.psha_adapter import OpenquakeSimplePshaAdapter
+from nzshm_model.psha_adapter.openquake import OpenquakeSimplePshaAdapter
 from nzshm_model.source_logic_tree.version2 import SourceLogicTree
 
 FIXTURE_PATH = Path(__file__).parent.parent.parent / "fixtures"
